+++ src/mock/mod.rs
@@ -1,2 +1,0 @@
-pub mod mock_role;
-pub mod mock_user_role;